restructured index-files
