+++ app.html
@@ -12,7 +12,7 @@
     <header>
         <h1>Hello World!</h1>
         <nav class="flex.container">
-            <ul class="nav-list"><li><button><a href="./index.html">Tillbaka</a></button></li></ul>
+            <ul class="nav-list"><li><button><a href="./index/index.html">Tillbaka</a></button></li></ul>
         </nav>
     </header>
     <footer>
