+++ index/index.html
@@ -7,7 +7,7 @@
     <title>GreenSway</title>
     <link href="https://fonts.googleapis.com/css2?family=Amatic+SC:wght@700&family=Tangerine:wght@700&display=swap" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css2?family=Amatic+SC:wght@700&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="./index/site.css">
+    <link rel="stylesheet" href="./site.css">
     
 </head>
 <body>
@@ -19,12 +19,12 @@
                 <li><button id="plan">Affärsplan</button></li>
                 <li><button id="idea">Produktidé</button></li>
                 <li><button id="contact">Kontakt</button></li>
-                <li><button><a href="./app.html">App</a></button></li>
+                <li><button><a href="../app.html">App</a></button></li>
             </ul>
         </nav>
     </header>
     <main id="main">
-        <img id="intro" src="./index/images/pexels-pixabay-255441.jpg" alt="Heart-shaped hole in a lush shrub.">
+        <img id="intro" src="./images/pexels-pixabay-255441.jpg" alt="Heart-shaped hole in a lush shrub.">
     </main>
     <footer> 
         <p id="copy"><span> "WiviWonderWoman" </span><br/>&copy; Wivianne Grapenholt</p>
