+++ index/product-idea.js
@@ -1,9 +1,9 @@
-import { main } from "../index.js";
+import { main } from "./index.js";
 
 export function showProductIdea() {
 
-    const image1 = './index/images/pexels-markus-spiske-1089438.jpg';
-    const image2 = './index/images/pexels-ready-made-3850452.jpg';
+    const image1 = './images/pexels-markus-spiske-1089438.jpg';
+    const image2 = './images/pexels-ready-made-3850452.jpg';
 
     main.innerHTML = `<div><h1>Produktidé</h1></div> 
     <div class="flex.container">
